+++ satellite/src/main/java/com/linbit/linstor/core/ControllerPeerConnector.java
@@ -4,6 +4,9 @@ import com.linbit.linstor.core.identifier.NodeName;
 import com.linbit.linstor.core.objects.Node;
 import com.linbit.linstor.netcom.Peer;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
 import java.util.UUID;
 
 public interface ControllerPeerConnector
@@ -12,7 +15,20 @@ public interface ControllerPeerConnector
 
     Peer getControllerPeer();
 
-    void setControllerPeer(UUID ctrlUuidRef, Peer controllerPeerRef, UUID nodeUuid, String nodeName);
+    /**
+     * Sets the controller peer object used later for replies to the controller.
+     *
+     * @param ctrlUuidRef
+     * @param controllerPeerRef
+     * @param nodeUuid
+     * @param nodeName
+     */
+    void setControllerPeer(
+        @Nullable UUID ctrlUuidRef,
+        @Nonnull Peer controllerPeerRef,
+        @Nonnull UUID nodeUuid,
+        @Nonnull String nodeName
+    );
 
     void setControllerPeerToCurrentLocalNode();