+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed version mismatch check between new satellites and a pre v1.25.0 controller
+- Fixed "node restore" to try to (re-) create tiebreaker resources if needed
 
 ## [1.25.0] - 2023-10-25