+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed "node restore" to try to (re-) create tiebreaker resources if needed
 - Fixed "node evacuate" no longer keeps diskless resource as a tiebreaker (in evacuating, but never deleting state)
 - Fixed "sp l" showing spurious warning that a diskless SP is offline (due to caching bug)
+- Fixed potential null pointer exception in AutoDiskfulTask
 
 ## [1.25.0] - 2023-10-25