+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed "node evacuate" no longer keeps diskless resource as a tiebreaker (in evacuating, but never deleting state)
 - Fixed "sp l" showing spurious warning that a diskless SP is offline (due to caching bug)
 - Fixed potential null pointer exception in AutoDiskfulTask
+- Fixed "r c --drbd-diskless" falsely requiring free peer-slot
 
 ## [1.25.0] - 2023-10-25